backend/requirements.txt
Flask
pymongo
requests
Flask-CORS
redis
kafka-python
PyJWT